Rewriting decoder buffering
Decoder had several bugs in its buffering logic that caused it to report
invalid data on redis data that was actually valid.

In particular the "second attempt to read" check incorrectly returns
error when the data is larger than the buffer size and the buffer is
initially empty.

This also fixes the (separate, but related) error reported in
gosexy/redis:
https://github.com/gosexy/redis/issues/42

This change rewrites the buffering logic to use standard bufio APIs
instead of reimplementing buffering logic in this library.